Move pages spec files to respective pages folders
