+++ client/src/js/pages/homePage/homePage.spec.js
@@ -1,4 +1,4 @@
-const HomePage = require('../../src/js/pages/homePage/homePage');
+const HomePage = require('./homePage');
 const StorageHub = require('watch-framework').StorageHub;
 const AudioHub = require('watch-framework').AudioHub;
